update: pages - moved file locations
