+++ policy-tool-frontend-v2/src/pages/Home.tsx
@@ -1,0 +1,22 @@
+import { Button, Grid, Typography } from '@material-ui/core'
+import { Link } from 'react-router-dom'
+
+export const Home: React.FC = () => (
+  <Grid container spacing={8}>
+    <Grid container item justify={'center'}>
+      <Typography variant={'h2'}>Welcome to the Policy Builder</Typography>
+    </Grid>
+    <Grid container justify={'center'}>
+      <Grid item>
+        <Button
+          variant={'outlined'}
+          color={'primary'}
+          component={Link}
+          to={'/builder'}
+        >
+          Create from scratch
+        </Button>
+      </Grid>
+    </Grid>
+  </Grid>
+)
