--- policy-tool-frontend-v2/src/pages/Home/index.ts
@@ -1,1 +1,0 @@
-export * from './Home'
