--- policy-tool-frontend-v2/src/pages/NotFound/index.ts
@@ -1,1 +1,0 @@
-export * from './NotFound'
